[AIRFLOW-5735] Move airflow.contrib.operators.s3_to_gcs_operator to core
