+++ UPDATING.md
@@ -364,6 +364,7 @@ The following table shows changes in import paths.
 |airflow.contrib.operators.pubsub_operator.PubSubSubscriptionDeleteOperator                                        |airflow.gcp.operators.pubsub.PubSubSubscriptionDeleteOperator                                              |
 |airflow.contrib.operators.pubsub_operator.PubSubTopicCreateOperator                                               |airflow.gcp.operators.pubsub.PubSubTopicCreateOperator                                                     |
 |airflow.contrib.operators.pubsub_operator.PubSubTopicDeleteOperator                                               |airflow.gcp.operators.pubsub.PubSubTopicDeleteOperator                                                     |
+|airflow.contrib.operators.s3_to_gcs_operator.S3ToGoogleCloudStorageOperator                                       |airflow.operators.s3_to_gcs_operator.S3ToGoogleCloudStorageOperator                                        |
 |airflow.contrib.operators.sql_to_gcs.BaseSQLToGoogleCloudStorageOperator                                          |airflow.operators.sql_to_gcs.BaseSQLToGoogleCloudStorageOperator                                           |
 |airflow.contrib.sensors.bigquery_sensor.BigQueryTableSensor                                                       |airflow.gcp.sensors.bigquery.BigQueryTableSensor                                                           |
 |airflow.contrib.sensors.gcp_transfer_sensor.GCPTransferServiceWaitForJobStatusSensor                              |airflow.gcp.sensors.cloud_storage_transfer_service.GCPTransferServiceWaitForJobStatusSensor                |
