+++ scripts/ci/pylint_todo.txt
@@ -62,7 +62,7 @@
 ./airflow/contrib/operators/s3_copy_object_operator.py
 ./airflow/contrib/operators/s3_delete_objects_operator.py
 ./airflow/contrib/operators/s3_list_operator.py
-./airflow/contrib/operators/s3_to_gcs_operator.py
+./airflow/operators/s3_to_gcs_operator.py
 ./airflow/contrib/operators/s3_to_sftp_operator.py
 ./airflow/contrib/operators/sagemaker_base_operator.py
 ./airflow/contrib/operators/sagemaker_endpoint_config_operator.py
