+++ tests/operators/test_s3_to_gcs_operator.py
@@ -19,7 +19,7 @@
 
 import unittest
 
-from airflow.contrib.operators.s3_to_gcs_operator import S3ToGoogleCloudStorageOperator
+from airflow.operators.s3_to_gcs_operator import S3ToGoogleCloudStorageOperator
 from tests.compat import mock
 
 TASK_ID = 'test-s3-gcs-operator'
@@ -51,10 +51,10 @@ class TestS3ToGoogleCloudStorageOperator(unittest.TestCase):
         self.assertEqual(operator.gcp_conn_id, GCS_CONN_ID)
         self.assertEqual(operator.dest_gcs, GCS_PATH_PREFIX)
 
-    @mock.patch('airflow.contrib.operators.s3_to_gcs_operator.S3Hook')
+    @mock.patch('airflow.operators.s3_to_gcs_operator.S3Hook')
     @mock.patch('airflow.contrib.operators.s3_list_operator.S3Hook')
     @mock.patch(
-        'airflow.contrib.operators.s3_to_gcs_operator.GoogleCloudStorageHook')
+        'airflow.operators.s3_to_gcs_operator.GoogleCloudStorageHook')
     def test_execute(self, gcs_mock_hook, s3_one_mock_hook, s3_two_mock_hook):
         """Test the execute function when the run is successful."""
 
@@ -86,10 +86,10 @@ class TestS3ToGoogleCloudStorageOperator(unittest.TestCase):
         # we expect MOCK_FILES to be uploaded
         self.assertEqual(sorted(MOCK_FILES), sorted(uploaded_files))
 
-    @mock.patch('airflow.contrib.operators.s3_to_gcs_operator.S3Hook')
+    @mock.patch('airflow.operators.s3_to_gcs_operator.S3Hook')
     @mock.patch('airflow.contrib.operators.s3_list_operator.S3Hook')
     @mock.patch(
-        'airflow.contrib.operators.s3_to_gcs_operator.GoogleCloudStorageHook')
+        'airflow.operators.s3_to_gcs_operator.GoogleCloudStorageHook')
     def test_execute_with_gzip(self, gcs_mock_hook, s3_one_mock_hook, s3_two_mock_hook):
         """Test the execute function when the run is successful."""
